// apps/web/src/lib/generated/models/DealRightType.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DealRightType = 'research' | 'development' | 'manufacturing' | 'commercialization' | 'co_development' | 'co_promotion' | 'distribution' | 'option' | 'other';